titleize instructors before importing from registrar API.
