+++ backend/tests/review/test_models.py
@@ -2,7 +2,7 @@ from django.test import TestCase
 
 from courses.models import Instructor
 from courses.util import get_or_create_course_and_section
-from review.import_utils.import_to_db import titleize
+from review.util import titleize
 from review.models import Review, ReviewBit
